# SentimentAnalyser/ClientApp/src/generated/api-client/model/sentimentRating.ts
/**
 * Sentiment Analyser API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SentimentRating = 0 | 4 | 8 | -8 | -5;

export const SentimentRating = {
    Modest: 0 as SentimentRating,
    Nice: 4 as SentimentRating,
    Excellent: 8 as SentimentRating,
    Horrible: -8 as SentimentRating,
    Ugly: -5 as SentimentRating
};